+++ common/config/test.php
@@ -89,7 +89,7 @@ return [
         ],
         'user' => [
             'class' => 'common\classes\WKUser',
-            'identityClass' => 'common\models\base\Person',
+            'identityClass' => 'domain\models\base\Person',
             'enableAutoLogin' => true,
             'identityCookie' => ['name' => '_identity-wk-portal', 'httpOnly' => true],
         ],